+++ Diagram.SRC/Diagram/Src/Lib/Fonts.cs
@@ -52,7 +52,7 @@ namespace Diagram
             bool italic = false;
             bool underline = false;
             bool strikeout = false;
-            float fontSize = 12F;
+            double fontSize = 12F;
 
             foreach (XElement el in element.Descendants())
             {
@@ -117,7 +117,7 @@ namespace Diagram
                 fontStyle = fontStyle | FontStyle.Strikeout;
             }
 
-            Font font = new Font(fontName, fontSize, fontStyle);
+            Font font = new Font(fontName, (int)fontSize, fontStyle);
             return font;
         }
