+++ Diagram.SRC/Diagram/Src/Lib/Patterns.cs
@@ -104,14 +104,14 @@ namespace Diagram
 
         /// <summary>
         ///  parse integer (representing node id) from goto command string </summary>
-        public static int getGotoIdCommand(string text)
+        public static long getGotoIdCommand(string text)
         {
             Match match = Regex.Match(text, @"^\s*#(\\d+)\s*$", RegexOptions.IgnoreCase);
 
 	        if (match.Success)
 	        {
 	            string key = match.Groups[1].Value;
-                return Int32.Parse(key);
+                return Int64.Parse(key);
             }
 
             return -1;
